Remove provider and update README
