+++ README.md
@@ -7,7 +7,7 @@
 
 A Markdown renderer for Vapor and Leaf. This uses the [Vapor Markdown](https://github.com/vapor/markdown) package to wrap [cmark](https://github.com/jgm/cmark) (though a [fork](https://github.com/brokenhandsio/cmark-gfm) is used to make it work with Swift PM), so it understands [Common Mark](http://commonmark.org). A quick reference guide for Common Mark can be found [here](http://commonmark.org/help/). It also supports [Github Flavored Markdown](https://guides.github.com/features/mastering-markdown/).
 
-## Use
+## Usage
 
 Once set up, you can use it in your Leaf template files like any other tag:
 
@@ -38,12 +38,15 @@ Add Leaf Markdown as a dependency in your `Package.swift` file:
 
 ### Register with Leaf
 
-Alternatively, you can also directly add the Tag onto your `LeafRenderer` if desired. During your setup (for example, in `Droplet+Setup.swift`), register your tag as so:
+To add the tag to Leaf, add it to your `LeafTagConfig`:
 
 ```swift
-if let leaf = drop.view as? LeafRenderer {
-    leaf.stem.register(Markdown())
-}
+try services.register(LeafProvider())
+var tags = LeafTagConfig.default()
+tags.use(Markdown(), as: "markdown")
+services.register(tags)
 ```
 
-Don't forget to import MarkdownProvider in the file you register the tag or add the provider in with `import MarkdownProvider`
+**Note:** it's important that you register the `LeafProvider` first otherwise this will override your `LeafTagConfig`.
+
+Don't forget to import LeafMarkdown in the file you register the tag with `import LeafMarkdown`.
